requirements.txt
crewai
crewai_tools

load_dotenv
langchain-huggingface